3619: make regions and providers into shared constants
